Split setup command in more descipt files
